Beginning styling on finished components before completing remaining functionality.
The top and bottom navigations are now finished and working. I might revisit them to add in responsive code for tablet, mobile, etc. The color scheme overall could change, but it is designed in Boston Bruins colors as of now to give an idea of how the finished product might look.